autotools: delete unit test result files if the test passed
Test status at this phase still can be PASSED/FAILED.
In previous version, at this point of the script
we knew the test has PASSED.
This no longer the case.

+++ scripts/test-grep.sh
@@ -31,6 +31,8 @@ if egrep -q 'ERROR: (LeakSanitizer|AddressSanitizer)' $1.result; then
 fi
 
 # keep only result file if test failed
-rm -f -- $1.result
+if test $exit_status -eq 0 ; then
+  rm -f -- $1.result
+fi
 
 exit $exit_status
